update user story 1

+++ domain/src/main/scala/model/User.scala
@@ -6,4 +6,4 @@ case class UserStatus(value: String)
 case class Department(name: String)
 
 case class User(userId: UserId, email: String, password: String, entryCompanyDate: String,
-                userRole: UserRole, department: Department, annualLeave: Long, userStatus: UserStatus)
+                userRole: UserRole, department: Department, annualLeave: Float, userStatus: UserStatus)
